Spring security

+++ src/main/java/pl/sk8/trickslist/trick/TrickRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface TrickRepository extends JpaRepository<Trick, Long> {
+public interface TrickRepository extends JpaRepository<Trick, Integer> {
     List<Trick> findByName(String name);
     Optional<Trick> findById(int id);
